Address additional PR review feedback: reduce code duplication and improve maintainability
- Extract date comparison logic into shared isMigrationDate() utility function
- Move getHubSpotDealStatus() and getACDealStatus() to shared migration-constants.js
- Remove duplicate methods from both update-hubspot-close-dates.js and data-gap-analyzer.js
- Create getDefaultFields() helper in CSV reporter to reduce field duplication
- Fix hardcoded date in error message to use dynamic migration date constant
- Improve code maintainability by centralizing migration-related utilities

Addresses latest GitHub Copilot PR review comments on code duplication and consistency
