+++ config/migration-constants.js
@@ -22,5 +22,31 @@ module.exports = {
       '2': 'lost',
       '3': 'open' // in progress
     }
+  },
+
+  // Utility functions for migration date checking
+  isMigrationDate(dateString) {
+    if (!dateString) return false;
+    return new Date(dateString).toISOString().split('T')[0] === this.MIGRATION_DATE;
+  },
+
+  // Deal status helper functions
+  getHubSpotDealStatus(stage) {
+    if (!stage) return 'unknown';
+    
+    const lowerStage = stage.toLowerCase();
+    if (lowerStage.includes('closedwon') || lowerStage.includes('won')) return 'won';
+    if (lowerStage.includes('closedlost') || lowerStage.includes('lost')) return 'lost';
+    return 'open';
+  },
+
+  getACDealStatus(status) {
+    switch (status) {
+      case '0': return 'open';
+      case '1': return 'won';
+      case '2': return 'lost';
+      case '3': return 'open'; // in progress
+      default: return 'unknown';
+    }
   }
 };
